requirements.txt
fastapi
uvicorn[standard]

SQLAlchemy==2.0.32
psycopg==3.2.12
alembic==1.17.0

python-dotenv==1.0.1
pydantic
pydantic-settings==2.10.1

passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
python-multipart==0.0.20
email-validator==2.2.0

orjson

pytest==8.4.1
httpx
